requirements.txt
asgiref==3.8.1
boto3==1.34.147
botocore==1.34.147
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.5.0
django-crispy-forms==2.3
django-heroku==0.3.1
django-sslserver==0.22
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==22.0.0
jmespath==1.0.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
